% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{dashboard}
\alias{dashboard}
\title{Shiny Dashboard to explore the California flights dataset}
\usage{
dashboard()
}
\value{
Dashboard of the dataset
}
\description{
Shiny Dashboard to explore the California flights dataset
}
