% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights.R
\docType{data}
\name{flights}
\alias{flights}
\title{Flights data}
\format{Data frame with columns
\describe{
\item{year, month, day}{Date of departure.}
\item{dep_time, arr_time}{Actual departure and arrival times (format HHMM or HMM), local tz.}
\item{dep_delay, arr_delay}{Departure and arrival delays, in minutes.
  Negative times represent early departures/arrivals.}
\item{carrier}{Two letter carrier abbreviation. See \code{\link{airlines}}
  to get name.}
\item{flight}{Flight number.}
\item{tailnum}{Plane tail number. See \code{planes} for additional metadata.}
\item{origin, dest}{Origin and destination. See \code{\link{airports}} for
  additional metadata.}
\item{distance}{Distance between airports, in miles.}
}}
\source{
RITA, Bureau of transportation statistics,
 \url{https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236&DB_Short_Name=On-Time}
}
\usage{
flights
}
\description{
On-time data for all flights that departed CA (i.e. LAX, SAN, SFO, PSP, SNA, OAK, ONT, SJC, ACV, BFL, BUR, CEC, CLD, FAT, LGB, MMH, MRY, RDD,SBA,SMF or SMX) in
2015.
}
\keyword{datasets}
