% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDashboard.R
\name{runDashboard}
\alias{runDashboard}
\title{California Flights Dashboard}
\usage{
runDashboard()
}
\value{

}
\description{
Title
}
\examples{
## Only run this example in interactive R sessions
if(interactive()){
   runDashboard()
}
}
