% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDashboard.R
\name{runDashboard}
\alias{runDashboard}
\title{CA flight Dashboard}
\usage{
runDashboard()
}
\value{

}
\description{
This dashboard explore the datasets in the \code{caflights} package. It shows all flights that departed from California to various destinations in the United States, Puerto Rico, and the American Virgin Islands) in 2015. You can find the dashboard repository from this link \url{https://github.com/denironyx/caflights15}
}
\examples{
## Only run this example in interactive R sessions
if(interactive()){
   runDashboard()
}
}
